% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{flowdir2ldd}
\alias{flowdir2ldd}
\title{Reclassifies  flow-direction grid to a ldd}
\usage{
flowdir2ldd(flowdir, noflow = 0)
}
\arguments{
\item{flowdir}{Input flow-direction map}

\item{noflow}{a value defined for sinks (default = 0)}
}
\value{
a ldd \code{RasterLayer} object
}
\description{
This function converts the flow-direction raster (by ArcGIS) to a ldd (by PCRaster).
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
