% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_functions.R
\name{getCellsGrid}
\alias{getCellsGrid}
\title{getCellsGrid}
\usage{
getCellsGrid(r)
}
\arguments{
\item{r}{input RasterLayer}
}
\value{
a RasterLayer with cell ids assigned in each cell
}
\description{
Assigns the cell id into each RasterLayer cell
}
