% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{getInlets}
\alias{getInlets}
\title{Identify inlets into desalinated spatial units}
\usage{
getInlets(r, network, ldd)
}
\arguments{
\item{r}{a spatial unit \code{RasterLayer} object}

\item{network}{a stream-network \code{RasterLayer} object}

\item{ldd}{a ldd \code{RasterLayer} object}
}
\value{
a named \code{list} containing a inlet \code{RasterLayer} and \code{data.frame} objects
}
\description{
Utilize the ldd and the stream-network to identify the inlets into a deliniated spatial units, e.g., river basins or administrative units.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
