% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{getUpstream}
\alias{getUpstream}
\title{Get upstream cell indices}
\usage{
getUpstream(x)
}
\arguments{
\item{x}{a ldd \code{RasterLayer} object}
}
\value{
a \code{data.frame} associating the id of each grid cell with its one or more direct upstream cells
}
\description{
Identify the upstream cell flowing into each grid cell, according to a ldd.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
