% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_functions.R
\name{na_to_val}
\alias{na_to_val}
\title{na_to_val}
\usage{
na_to_val(r, val = 0)
}
\arguments{
\item{r}{input data object (e.g., RasterLayer)}

\item{val}{alternative value to be assigned instead of NA (Default: 0)}
}
\value{
the data object with all NAs replaced by the assigned value
}
\description{
The function takes a data object, e.g., matrix, RasterLayer, or a
vector and convert NA values to a selected value (default is  0)
}
