% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{ncdfInfo}
\alias{ncdfInfo}
\title{Gets Information and metadata from netcdf}
\usage{
ncdfInfo(pth, dim = TRUE, attrs = FALSE)
}
\arguments{
\item{pth}{a path to a NetCDF file with the extension '.nc' or '.nc4'}

\item{dim}{\code{logical} if TRUE, dimensions the files are returned}

\item{attrs}{\code{logical} if TRUE, variable and global attributes are returned}
}
\value{
a named list
}
\description{
This function extracts dimensions and attributes from 'NetCDF' file.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
