% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handling_functions.R
\name{val_to_na}
\alias{val_to_na}
\title{val_to_na}
\usage{
val_to_na(r, val = 0)
}
\arguments{
\item{r}{input data object (e.g., RasterLayer)}

\item{val}{selected value to be replaced by NA (Default: 0)}
}
\value{
the data object with all selected values replaced by NAs
}
\description{
The function takes a data object, e.g., matrix, RasterLayer, or a
vector and convert a selected value (default is  0) to NA
}
