% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{zoomClip}
\alias{zoomClip}
\title{Point & click to zoom in on a RasterLayer object}
\usage{
zoomClip(r, size = 15)
}
\arguments{
\item{r}{Input flow direction map (Only PCRaster ldd is accepted)}

\item{size}{pre-defined distance (grid-cell) for defining a zoom-in window (default = 15)}
}
\value{
a \code{RasterLayer} object
}
\description{
This function zoom in to a RasterLayer using R \code{base} graphics and the \code{raster} libraries.
Calling the function plots the raster layer. Zoom in is triggered by a user point and click.
The scope of the output raster is defined by the \code{size} argument, resulting in a window of area \code{(2*size + 1)^2}.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
