% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{getInlets}
\alias{getInlets}
\title{Identify inlets into desalinated spatial units}
\usage{
getInlets(r, network, ldd, plot_ = TRUE)
}
\arguments{
\item{r}{a spatial unit \code{RasterLayer} object}

\item{network}{a stream-network \code{RasterLayer} object}

\item{ldd}{a ldd \code{RasterLayer} object}

\item{plot}{a logical variable. If \code{TRUE}, a plot of the inlet points, spatial units, and river network is plotted \code{RasterLayer} object}
}
\value{
a \code{data.frame} object with the inflow point location, cell ids, and associated from- and to-zones
}
\description{
Utilize the ldd and the stream-network to identify the inlets into a delineated spatial units, e.g., river basins or administrative units.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
