% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoBasins_functions.R
\name{maskmapFromXY}
\alias{maskmapFromXY}
\title{Delineate a watershed based on ldd and an outlet point}
\usage{
maskmapFromXY(ldd, lat, lon, loud = TRUE)
}
\arguments{
\item{ldd}{Input flow direction map (Only PCRaster ldd is accepted)}

\item{lat, lon}{latitude and longitude of the outlet in the ldd reference system}

\item{loud}{\code{logical} if TRUE, the search list (including cell ids) will be printed every iteration}
}
\value{
a named list
}
\description{
This function create basin mask map based on user inputs of ldd and XY location of the outlet.
}
\examples{
\dontrun{
PUT EXAMPLE HERE
}
}
